Moved `ContainerResultMatcher` to a separate file
